refactor: 변수 및 네이밍 통일 & 파일 정리
