+++ Expo1900/Expo1900/Controller/ExhibitionWorkListViewController.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class ExhibitionWorkListViewController: UIViewController {
+final class ExhibitionWorkListViewController: UIViewController {
     
     let exhibitionWorkTableView: UITableView = UITableView()
     
@@ -21,24 +21,7 @@ class ExhibitionWorkListViewController: UIViewController {
         exhibitionWorkTableView.delegate = self
     }
     
-    func setUpTableView() {
-        self.exhibitionWorkTableView.register(TableViewCell.self, forCellReuseIdentifier: TableViewCell.cellIdentifier)
-        self.view.addSubview(self.exhibitionWorkTableView)
-        
-        self.exhibitionWorkTableView.translatesAutoresizingMaskIntoConstraints = false
-        self.view.addConstraint(NSLayoutConstraint(item: self.exhibitionWorkTableView,
-                                                   attribute: .top, relatedBy: .equal, toItem: self.view, attribute: .top,
-                                                   multiplier: 1.0, constant: 0))
-        self.view.addConstraint(NSLayoutConstraint(item: self.exhibitionWorkTableView,
-                                                   attribute: .bottom, relatedBy: .equal, toItem: self.view,
-                                                   attribute: .bottom, multiplier: 1.0, constant: 0))
-        self.view.addConstraint(NSLayoutConstraint(item: self.exhibitionWorkTableView,
-                                                   attribute: .leading, relatedBy: .equal, toItem: self.view,
-                                                   attribute: .leading, multiplier: 1.0, constant: 0))
-        self.view.addConstraint(NSLayoutConstraint(item: self.exhibitionWorkTableView,
-                                                   attribute: .trailing, relatedBy: .equal, toItem: self.view,
-                                                   attribute: .trailing, multiplier: 1.0, constant: 0))
-    }
+    
     
     func setUpJSONData() {
         let jsonDecoder: JSONDecoder = JSONDecoder()
